hoan thanh chuc nang ngay 1

+++ src/main/java/com/ra/hn_jv231229_md03_watchfilmonline_project/service/design/IFilmService.java
@@ -17,7 +17,7 @@ public interface IFilmService
 
     List<Film> searchFilmRelative(String infoToSearch, String columnName);
 
-    Boolean deleteFilmById(int id);
+    Boolean deleteFilmById(long id);
 
     Integer countNumberOfFilms();
 }
